# public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZBY Taxi Dispatch & Driver API",
    "description": "Rasmiy ZBY Taxi Backend REST API hujjati. Flutter mobil ilovasi (Haydovchi va Mijoz) hamda Dispetcherlik veb-paneli uchun to'liq API spetsifikatsiyasi.",
    "version": "1.0.0",
    "contact": {
      "name": "ZBY Taxi Engineering Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Mahalliy ishlab chiqish serveri (Local Dev)"
    },
    {
      "url": "http://10.0.2.2:8000/api",
      "description": "Android Emulator orqali localhostga ulanish"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token",
        "description": "Tizimga kirishda olingan Sanctum tokeni. Format: 'Bearer <token>'"
      }
    },
...
  },
  "paths": {
...
    "/telegram/webhook": {
      "post": {
        "tags": ["5. Telegram Bot Integratsiyasi"],
        "summary": "Telegram Bot Webhook Update Handler",
        "description": "Telegram Bot serveridan keladigan barcha xabarlar, skrinshot rasmlari va /start komandalarini qayta ishlovchi endpoint.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Update qabul qilindi" }
        }
      }
    },
...
  }
}